requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
requests
pytest
httpx
pypdf
sentence-transformers
faiss-cpu
google-generativeai
